细节调整

+++ app/src/main/java/com/loyo/oa/v2/activityui/signinnew/model/AudioModel.java
@@ -6,5 +6,5 @@ package com.loyo.oa.v2.activityui.signinnew.model;
 
 public class AudioModel {
     public String url;
-    public int length;
+    public long length;
 }
